fixes associated with parsing very large --inhex=FN files; sg_logs: start on seagate farm
git-svn-id: https://svn.bingwo.ca/repos/sg3_utils/trunk@1015 6180dd3e-e324-4e3e-922d-17de1ae2f315

+++ lib/sg_pr2serr.c
@@ -485,7 +485,7 @@ sgj_pr_hr(sgj_state * jsp, const char * fmt, ...)
 
         va_start(args, fmt);
         ln = vsnprintf(b, blen, fmt, args);
-        if ((ln > 0) && (ln < blen)) {
+        if ((ln > 0) && (ln < (size_t)blen)) {
             char * cp;
 
              /* deal with leading, trailing and embedded newlines */
